feat: Enhance Wi-Fi Manager with RSSI, Uptime, and Peer List Management
- Added functions to retrieve RSSI, Wi-Fi uptime, and current channel.
- Implemented peer list management with functions to set and load peers from NVS.
- Updated factory reset to clear peer list.
- Initialized peer list loading during Wi-Fi manager initialization.

refactor: Remove WebSocket Client Implementation

- Deleted ws_client.cpp as it is no longer needed in the project.

feat: Implement WebSocket Server with Client Management

- Introduced WebSocket server functionality with client connection management.
- Added support for handling incoming messages and broadcasting to clients.
- Implemented status and heartbeat acknowledgment messages.

feat: Add ESP-NOW Communication Layer

- Created espnow_comm.h and espnow_comm.cpp for handling ESP-NOW messaging.
- Implemented functions for peer management, message sending, and receiving.
- Added support for broadcasting messages and managing peers from a CSV list.

+++ src/utils.cpp
@@ -1,6 +1,6 @@
 #include "utils.h"
 
-std::string toBinaryString(uint16_t value, int bits)
+std::string toBinaryString(uint32_t value, int bits)
 {
     std::string binary = "";
     for (int i = bits - 1; i >= 0; i--)
